Convert StatusBarNotifications into NotificationGroups. Design doc: go/aae-notification-grouping
This is the first step implementing the grouping functionality in notification center. With this CL, StatusBarNotifications from the same package are grouped and stored in the NotificationGroup object. No presentation layer is implemented in this CL: only the group header notification in a group is shown. When no group header notification is available, the first (most recent) child notification is shown. The view (inbox style notification template, and the list view of child notifications) and interaction (expansion/collapse/swipe) will be implemented in the follow-up CLs.

Bug: 80268960
Test: Use KitchenSink app to send notifications
Change-Id: I8114b73a1721bc0b1652f4f0e7bc984ae9e838b2