Roll forward "Add clear all button to the notification center."
The CarSystemUI builds the CarNotification as a library and has a different layout for the UI, but it doesn't have an app bar so a clear all button cannot be put there. Therefore the clear all button can be null, hence added a null check in case CarSystemUI mulfunctions.

This reverts commit 6f680a1ccdd954823ea211f0f46381103e775262.

Reason for revert: rolling forward after null check

Test: Open notification center and tap clear all
Change-Id: I731ab4ba79f9721ec7a39679349b47ef1adb2be9