Revert "Add a custom FrameLayout to show Heads up notification view that will disable touch for 700 ms."
This reverts commit e734973092d072d11de30e637b834393a4bee7f6.

Reason for revert: This approach in not 100% correct. We should allow the user to swipe off the Notifications in 700 ms and only stop the accidental clicks on Action button and car that can open the Intent. 
Right now each time the view is inflated 700 ms gets added. It happens for ongoing notifications which sends update like maps.

Change-Id: I8fef7d2f6e6dd65bb1f74059325526feaca1e409