Enlarge the group expand/collapse touch target.
- When collapsed, the entire group notification card is clickable to expand a group
- When expanded, only the group header is clickable to collapse a group

Reason the click logic is in the group view holder instead of the group summary view holder:
The group summary notification can be any normal notification; It does not have to be an auto generated group summary. Therefore, the click logic has to be in the group notification view. I tried to make clicks fall through on the recycler view but it turned out to be difficult, so I just added a transparent view on top that intercepts all clicks.

Bug: 128344300
Test: Send 4+ notifications from Kitchensink and test touch target
Change-Id: Ia9ee70948f7fc9dcbe70ee33a54461a619995b2d