Fix for jank when dismissing a notification group from NC.
When a notification group is dismissed, the platform raises removed
event for each notifications in quick succession. (<5 ms) This causes RV
to update for every item in the group, resulting in Jank. This change
adds short delay before triggering UI update and removes queued update
if another change is requested before the update is triggered.

Bug: 132902463
Test: Manual Test.
Change-Id: I600f71e9917157035da59d5159e90bf4427ed3a0