Revert "Use DiffUtil instead of notifyDataSetChanged."
This reverts commit 832c236f615744ca5dfef944a2a5c0cdcaf2ee24.

Reason for revert: This change is causing regression issues. There are certain cases thats causing the UI to break For example, when we try to dismiss any message within the group other than the first one it throws an error. This was solving a extreme edge case where when there is a ongoing update as in downloads, then notifyDataSetChange was rebinding all the notifications. We can deal with this in R given the time we have.

Test: Robolectric and manual
Change-Id: I49496d6d303ef713d9dbf24dafc821da32da5399